pyproject.toml
[project]
name = "code-clone-detector"
version = "0.1.0"
description = "JavaScript Code Clone Detection using SAGA"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
